requirements.txt
asyncpg==0.22.0
asyncpg==0.26.0
git+https://github.com/Rapptz/discord-ext-menus
git+https://github.com/oliver-ni/discord-ext-menus-views
flake8==3.9.2
discord.py==2.0.0
pytz==2021.1
python-dotenv==0.15.0
